Version 2.0
Implemented "BADCOUNTRY":
If UpChecker determines that the device is online, it will check if the IP's country matches the determined BADCOUNTRY. If it does, it will notify the user via Telegram and reboot the system. This is intended to fix VPN outages. If you wish to use it, you should set BADCOUNTRY to whatever country your true IP is (If you're in the USA, BADCOUNTRY = 'US') and set the VPN to a different country. If you don't wish to use this function, just set BADCOUNTRY to any country, that is not your own (e.g. If you are located in the US, set BADCOUNTRY to GB)