Offline behavior improved
There was a pretty big issue where the device going offline wouldn't trigger the intended actions because it would try to notify the user via telegram, which would not work (because offline) which would lead to the script failing and nothing further happening. Now there's more failsaves and I implemented a lot of "try" and "except" code which logs and reboots when the telegram message sending fails.